Revert "Refactor dita function from alarms_parser"
